+++ crates/block/hadris-fat/src/async_send.rs
@@ -1,0 +1,4 @@
+#[allow(unused_macros)]
+macro_rules! io_transform {
+    ($($item:tt)*) => { hadris_macros::send_async! { $($item)* } };
+}