+++ crates/core/hadris-fs/tests/common/mod.rs
@@ -52,11 +52,14 @@ pub fn block_on<F: core::future::Future>(future: F) -> F::Output {
     }
 }
 
+#[cfg(feature = "sync")]
 #[path = "mode_sync.rs"]
 pub mod sync;
 
+#[cfg(feature = "async")]
 #[path = "mode_asynch.rs"]
 pub mod asynch;
 
+#[cfg(feature = "async-send")]
 #[path = "mode_send.rs"]
 pub mod send;